requirements.txt
django>=5.0,<6.0
djangorestframework
django-filter
drf-spectacular
psycopg2-binary
dj-database-url
python-dotenv